Move FIPS E2E test to new compliance category
...rather than as part of the "dataplane" category so
it's not automatically run by 'subctl verify' as a
connectivity test.
